20.03.06) 답변이 있는 경우 삭제 실패 기능 구현
Question 객체에 notHasAnswers() 추가.
삭제 실패 시 질문 상세보기 페이지에 그대로 머물기.

+++ src/main/java/com/codessquad/qna/QuestionController.java
@@ -90,8 +90,12 @@ public class QuestionController {
         try {
             hasPermission(httpSession, writer);
             Question question = findQuestionById(questionRepository, id);
-            question.delete();
-            questionRepository.save(question);
+            if (question.notHasAnswers()) {
+                question.delete();
+                questionRepository.save(question);
+                return "redirect:/questions/{id}";
+            }
+            System.out.println("답변이 있는 경우 지울 수 없습니다.");
             return "redirect:/";
         } catch (IllegalStateException e) {
             model.addAttribute("errorMessage", e.getMessage());
